% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_custom.R
\name{signif_custom}
\alias{signif_custom}
\title{Rounding of numbers to significant digits with options}
\usage{
signif_custom(x, digits = 1, option = ceiling)
}
\arguments{
\item{x}{Numeric vector to be rounded to set significant figures}

\item{digits}{Number of significant figures to round to}

\item{option}{Specific rounding function to apply; options are \code{ceiling},
\code{floor}, \code{trunc}, or \code{round}}
}
\value{
Numeric vector of rounded numbers
}
\description{
A combination of the the \code{\link[base]{signif}()} and
\code{\link[base]{ceiling}()}/\code{\link[base]{floor}()} functions. It
allows numbers to be rounded to a provided number of significant digits, but
to specifically have \code{\link[base]{ceiling}()} or
\code{\link[base]{floor}()} applied for the rounding.
}
\examples{
signif_custom(1014, digits = 2)
signif_custom(67362, digits = 3, option = floor)
}
