% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_scientific.R
\name{to_scientific}
\alias{to_scientific}
\title{Convert numbers into formatted scientific notation expressions}
\usage{
to_scientific(
  x,
  digits = 2,
  max_cut = 10^5,
  min_cut = 10^-3,
  common = FALSE,
  factor = NULL,
  trailing = TRUE,
  units = NULL
)
}
\arguments{
\item{x}{Numeric vector, gives numbers to convert to scientific notation}

\item{digits}{Number of significant figures to display in the converted
output}

\item{max_cut}{Above this number, values will be converted into scientific
format (inclusive)}

\item{min_cut}{Below this number, values will be converted into scientific
format (exclusive)}

\item{common}{Should the formatted numbers share a common exponential factor?}

\item{factor}{Provides a value for the common exponential factor; if \code{NULL},
the most occurring exponent will be used}

\item{trailing}{Should trailing zeros be included in the formatted number?}

\item{units}{If provided, adds units to the converted output}
}
\value{
An expression containing the input numbers in scientific format
}
\description{
This function converts a vector of numbers to formatted scientific notation
expressions. It is called by \code{\link{label_sci}()} to convert the
\code{\link[ggplot2]{ggplot2}} scale function labels to scientific format.
}
\examples{
a <- c(NA, 10657, 0.00187, 0.0005, 102, 0)
to_scientific(a)
}
